changed array size parameters to size_t

+++ llama.h
@@ -928,12 +928,12 @@ extern "C" {
     LLAMA_API void llama_sample_dry(
           llama_token_data_array * candidates,
                const llama_token * last_tokens,
-                             int   last_tokens_size,
+                          size_t   last_tokens_size,
                            float   dry_base,
                            float   dry_multiplier,
                              int   dry_allowed_length,
                const llama_token * dry_seq_breakers,
-                             int   dry_seq_breakers_size);
+                          size_t   dry_seq_breakers_size);
 
     /// @details Tail Free Sampling described in https://www.trentonbricken.com/Tail-Free-Sampling/.
     LLAMA_API void llama_sample_tail_free(
